requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
asgiref==3.8.1
attrs==23.2.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
coverage==7.5.0
cycler==0.12.1
Cython==3.0.9
Django==5.0.3
django-filter==24.2
djangorestframework==3.15.1
Faker==24.4.0
Flask==3.0.2
fonttools==4.43.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
Mako==1.3.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.0
multidict==6.0.5
numpy==1.26.0
packaging==23.1
phonenumbers==8.13.36
Pillow==10.0.1
pluggy==1.3.0
progress==1.6
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.28
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.0.6
Werkzeug==3.0.1
yarl==1.9.4